# ios/CleaveAPI/Sources/CleaveAPI/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Cleave",
    "version": "0.1.0"
  },
  "paths": {
...
    "/expenses/{expense_id}/receipts": {
      "post": {
        "tags": [
          "receipts"
        ],
        "summary": "Upload Receipt",
        "operationId": "upload_receipt_expenses__expense_id__receipts_post",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "expense_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Expense Id"
            }
          }
        ],
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ReceiptResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "requestBody": {
          "content": {
            "application/octet-stream": {
              "schema": {
                "type": "string",
                "format": "binary"
              }
            }
          },
          "required": true
        }
      },
      "get": {
        "tags": [
          "receipts"
        ],
        "summary": "List Receipts",
        "operationId": "list_receipts_expenses__expense_id__receipts_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "expense_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Expense Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ReceiptResponse"
                  },
                  "title": "Response List Receipts Expenses  Expense Id  Receipts Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/transactions/{transaction_id}/receipts": {
      "post": {
        "tags": [
          "receipts"
        ],
        "summary": "Upload Transaction Receipt",
        "operationId": "upload_transaction_receipt_transactions__transaction_id__receipts_post",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "transaction_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Transaction Id"
            }
          }
        ],
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ReceiptResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "requestBody": {
          "content": {
            "application/octet-stream": {
              "schema": {
                "type": "string",
                "format": "binary"
              }
            }
          },
          "required": true
        }
      },
      "get": {
        "tags": [
          "receipts"
        ],
        "summary": "List Transaction Receipts",
        "operationId": "list_transaction_receipts_transactions__transaction_id__receipts_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "transaction_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Transaction Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ReceiptResponse"
                  },
                  "title": "Response List Transaction Receipts Transactions  Transaction Id  Receipts Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/receipts/{receipt_id}/content": {
      "get": {
        "tags": [
          "receipts"
        ],
        "summary": "Download Receipt",
        "operationId": "download_receipt_receipts__receipt_id__content_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "receipt_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Receipt Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/octet-stream": {
                "schema": {
                  "type": "string",
                  "format": "binary"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/receipts/{receipt_id}": {
      "delete": {
        "tags": [
          "receipts"
        ],
        "summary": "Delete Receipt",
        "operationId": "delete_receipt_receipts__receipt_id__delete",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "receipt_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Receipt Id"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/preferences": {
      "get": {
        "tags": [
          "preferences"
        ],
        "summary": "List Preferences",
        "description": "Every preference blob belonging to the caller (empty in open mode). Each is an opaque JSON string\nthe client decodes itself - e.g. the suggestion templates/decisions under key `suggestions.v1`. (The old\n`categories.v1` blob was retired in migration 0049; categories now sync via `/categories`.)",
        "operationId": "list_preferences_preferences_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/PreferenceResponse"
                  },
                  "type": "array",
                  "title": "Response List Preferences Preferences Get"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/preferences/{key}": {
      "put": {
        "tags": [
          "preferences"
        ],
        "summary": "Upsert Preference",
        "description": "Store (create or replace) the caller's preference blob under `key`. Scoped to `owner_identifier`,\nso users sharing a backend keep independent preferences.",
        "operationId": "upsert_preference_preferences__key__put",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "key",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Key"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PreferenceUpsert"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PreferenceResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/categories": {
      "get": {
        "tags": [
          "categories"
        ],
        "summary": "Get Categories",
        "description": "The caller's category taxonomy + raw\u2192canonical map (empty in open mode). `updated_at` is the newest\nrow timestamp - the client's last-write-wins watermark, replacing the old `categories.v1` blob's.",
        "operationId": "get_categories_categories_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CategoryConfig"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      },
      "put": {
        "tags": [
          "categories"
        ],
        "summary": "Put Categories",
        "description": "Idempotent owner-scoped replace-set: the caller's taxonomy + map become exactly the payload (rows\nabsent from it are deleted), in one transaction - mirroring the old whole-blob snapshot semantics\nrelationally. Device-authoritative; the app owns the set and pushes it after local edits.",
        "operationId": "put_categories_categories_put",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CategoryConfigUpsert"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CategoryConfig"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "CategoryConfig": {
        "properties": {
          "categories": {
            "items": {
              "$ref": "#/components/schemas/SpendCategoryItem"
            },
            "type": "array",
            "title": "Categories",
            "default": []
          },
          "maps": {
            "items": {
              "$ref": "#/components/schemas/CategoryMapItem"
            },
            "type": "array",
            "title": "Maps",
            "default": []
          },
          "updated_at": {
            "type": "string",
            "format": "date-time",
            "title": "Updated At"
          }
        },
        "type": "object",
        "title": "CategoryConfig",
        "description": "The caller's full category taxonomy + raw\u2192canonical map - the relational successor to the\n`categories.v1` blob. `updated_at` is the max row timestamp, the last-write-wins watermark the client\ncompares against its local sync watermark."
      },
      "CategoryConfigUpsert": {
        "properties": {
          "categories": {
            "items": {
              "$ref": "#/components/schemas/SpendCategoryItem"
            },
            "type": "array",
            "title": "Categories",
            "default": []
          },
          "maps": {
            "items": {
              "$ref": "#/components/schemas/CategoryMapItem"
            },
            "type": "array",
            "title": "Maps",
            "default": []
          }
        },
        "type": "object",
        "title": "CategoryConfigUpsert"
      },
      "CategoryMapItem": {
        "properties": {
          "raw_category": {
            "type": "string",
            "title": "Raw Category"
          },
          "canonical_category": {
            "type": "string",
            "title": "Canonical Category"
          },
          "source": {
            "type": "string",
            "title": "Source",
            "default": "manual"
          }
        },
        "type": "object",
        "required": [
          "raw_category",
          "canonical_category"
        ],
        "title": "CategoryMapItem"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "PreferenceResponse": {
        "properties": {
          "key": {
            "type": "string",
            "title": "Key"
          },
          "value": {
            "type": "string",
            "title": "Value"
          },
          "updated_at": {
            "type": "string",
            "format": "date-time",
            "title": "Updated At"
          }
        },
        "type": "object",
        "required": [
          "key",
          "value",
          "updated_at"
        ],
        "title": "PreferenceResponse"
      },
      "PreferenceUpsert": {
        "properties": {
          "value": {
            "type": "string",
            "title": "Value"
          }
        },
        "type": "object",
        "required": [
          "value"
        ],
        "title": "PreferenceUpsert"
      },
...
      "ReceiptResponse": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "expense_id": {
            "type": "string",
            "format": "uuid",
            "title": "Expense Id"
          },
          "transaction_id": {
            "type": "string",
            "format": "uuid",
            "title": "Transaction Id"
          },
          "bucket": {
            "type": "string",
            "title": "Bucket"
          },
          "object_key": {
            "type": "string",
            "title": "Object Key"
          },
          "content_type": {
            "type": "string",
            "title": "Content Type"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          }
        },
        "type": "object",
        "required": [
          "id",
          "bucket",
          "object_key",
          "created_at"
        ],
        "title": "ReceiptResponse"
      },
...
      "SpendCategoryItem": {
        "properties": {
          "name": {
            "type": "string",
            "title": "Name"
          },
          "icon": {
            "type": "string",
            "title": "Icon"
          },
          "color": {
            "type": "string",
            "title": "Color"
          },
          "position": {
            "type": "integer",
            "title": "Position",
            "default": 0
          },
          "builtin": {
            "type": "boolean",
            "title": "Builtin",
            "default": false
          }
        },
        "type": "object",
        "required": [
          "name"
        ],
        "title": "SpendCategoryItem"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
    },
    "securitySchemes": {
      "HTTPBearer": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  }
}